+++ libft/include/ft_printf.h
@@ -32,7 +32,7 @@ void			format_upper_hexadecimal(t_emitter *e, t_spec s,
 					unsigned int n);
 void			format_string(t_emitter *e, t_spec s, char *str);
 t_format_output	format_unsigned(t_spec s, unsigned long n, char *base);
-void			format_pointer(t_emitter *e, t_spec s, unsigned long long ptr);
+void			format_pointer(t_emitter *e, t_spec s, uintptr_t ptr);
 void			format_unsigned_integer(t_emitter *e, t_spec s, unsigned int n);
 
 #endif
